refactor(node): split into NodeBase and NodeDirect
