+++ src/tx/builder/field-types/ttl.ts
@@ -1,5 +1,5 @@
 import shortUInt from './short-u-int';
-import Node from '../../../Node';
+import Node from '../../../node/Direct';
 import { ArgumentError } from '../../../utils/errors';
 
 /**
