+++ src/tx/validator.ts
@@ -6,7 +6,8 @@ import { Tag, ConsensusProtocolVersion } from './builder/constants';
 import { buildTx, unpackTx } from './builder';
 import { concatBuffers, isAccountNotFoundError } from '../utils/other';
 import { Encoded, decode } from '../utils/encoder';
-import Node, { TransformNodeType } from '../Node';
+import { TransformNodeType } from '../node/Base';
+import Node from '../node/Direct';
 import { Account } from '../apis/node';
 import { genAggressiveCacheGetResponsesPolicy } from '../utils/autorest';
 import { UnexpectedTsError } from '../utils/errors';
